17 jun 2022

+++ lib/controller/payment_method_controller.dart
@@ -1,5 +1,4 @@
 import 'dart:ui';
-
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:ifreshoriginals_userapp/controller/order_controller.dart';
@@ -8,11 +7,10 @@ import 'package:square_in_app_payments/models.dart';
 import 'package:http/http.dart' as http;
 
 class PaymentMethodController extends GetxController {
-  double totalForPay = 0;
-  double deliveryCharge = 0;
+  int totalForPay = 0;
+  int deliveryCharge = 0;
 
   void startCardEntryFlowMethod(){
-
     InAppPayments.setSquareApplicationId("sandbox-sq0idb-DmyyChN1jRRS7sq8M3e_XQ");
     InAppPayments.startCardEntryFlow(
         onCardNonceRequestSuccess: onCardNonceRequestSuccess,
